# More efficient encoding of maps in interface files