--- a/src/full/Agda/TypeChecking/Serialise.hs
+++ b/src/full/Agda/TypeChecking/Serialise.hs
@@ -83,7 +83,7 @@ import Agda.Utils.Impossible
 -- 32-bit machines). Word64 does not have these problems.
 
 currentInterfaceVersion :: Word64
-currentInterfaceVersion = 20220628 * 10 + 0
+currentInterfaceVersion = 20220708 * 10 + 0
 
 -- | The result of 'encode' and 'encodeInterface'.